extended Tuple to a ComparableTuple so it can be used in a Histogram
renaming FilteringIterators to reflect what they are filtering (SAMRecord and VariantContext)
